+++ lib/src/models/sticker_kind.dart
@@ -1,16 +1,12 @@
 enum StickerKind {
   staticImage,
   gif,
-  lottie,
   unicode;
 
   static StickerKind fromExtension(String? ext) {
     switch ((ext ?? '').toLowerCase()) {
       case 'gif':
         return StickerKind.gif;
-      case 'json':
-      case 'lottie':
-        return StickerKind.lottie;
       default:
         return StickerKind.staticImage;
     }